moved plugins.ts to data folder
